COMMERCE-819 Move CPAttachmentFileEntryCreator out of internal
